refactor: export `extractFrameMetaTags` from root
