+++ site/pages/platforms/next.mdx
@@ -229,7 +229,7 @@ path of the website on Warpcast, they will be able to see the frame.
 
 ```tsx twoslash [app/page.tsx]
 // @noErrors
-import { extractFrameMetaTags } from 'frog/metadata'
+import { extractFrameMetaTags } from 'frog'
 import { type Metadata } from 'next'
 // The base URL of your app
 import { APP_URL } from './app-url'
