+++ src/index.ts
@@ -6,5 +6,6 @@ export {
   type ButtonResetProps,
 } from './components/Button.js'
 export { TextInput, type TextInputProps } from './components/TextInput.js'
+export { extractFrameMetaTags } from './utils/extractFrameMetaTags.js'
 
 export { Frog } from './frog.js'
